add ECN calculator for relative response factors
the relative response factors can be estimated for a set of compounds by
the effective carbon number. this classes the carbon atom types based on
first and second nearest neighbors, by identifying the functional groups
those carbon atoms belong to. each carbon atom class or functional group
is assigned a value. here the value is at least 0, since each
functional group is presumed to only reduce the effective carbon number
of the carbon it includes to 0. this is because functional group effect
on ionization isn't expected to go beyond the atoms in the functional
group.