separated smoothing and baseline subtract modules
baseline subtraction is separated from smoothing routines. the
motivation for smoothing for baseline subtraction might be to better
identify the starting and stopping points of the baseline curve, and
improving a fit of a linear or low order polynomial to baseline points.
indeed the asymmetric weights from the alternating least squares routine
can be used in a regression of a polynomial. that might be later
supported.

mostly the same thing is achieved by splining between the
points which are identified by curves, effectively assigning weight 0 to
the peak points and weight 1 to non-peak points.

also separated the creation of data from the sensitivity factors module