option to add window width to match
previously only the minimum distance from the interval center was used
to choose the match of an identified peak to. now the retention time
window is available to be used in the minimum. note the retention time
window is often dose dependent, so by default the window width is not
used.